[REFACTOR] extract non-test methods from OSTest into OSTestHarness superclass
